Fix Cuckoo Generator to handle reserved keywords
Allows for reserved keywords to be used as argument labels / parameters
in mocked types.

Fixes: https://github.com/Brightify/Cuckoo/issues/452

- Moves escaping util function to Utils and makes it internal since it
  now needs to be used outside of `Generator`.
- Uses escaping function in areas affected by the bug.
- Renames set of reserved names to reflect the larger scope (no longer
  used for just function names).
- Some minor clean-up to style and formatting on areas touched.

With this change you can now mock functions like:
https://developer.apple.com/documentation/foundation/filemanager/1407693-url
on `FileManager`.

Example of the problem:

Generate a mock for a protocol with the function:

```
func escape(for name: String) -> String
```

This would result in `for` as a function parameter used within the body
of the function for the mock and verification and stubbing proxies,
which would not compile. This change escapes keywords like this so the
generated code will compile.

Note: There was previous attempt at escaping reserved keywords
(https://github.com/Brightify/Cuckoo/pull/412), but it was incomplete and
only handled the case of a function name using a reserved keyword, which
seems to be less common since Apple's frameworks have many functions and
initializers that use `for` or `in` as an argument label.